Engine code refactoring (#399)
Bug fixes: 
* Fix transparency and texture issues
* Identified depth-texture bug that makes it unusable after rendering.
* Leave depth-buffer usable after rendering debug, ui and Xray objects (must be tested with bones)
* Fix bone rendering and depth-buffer display.
* [assimploader] Fix incomplete scenes loading.

Improvements:
* Texture API update and full code cleanup in the Engine part.
* Add GL_SCALAR define to match Scalar.
* Add automatic texture unit management.
* Remove option for openMP (use if available)

Documentation:
* Removed forgotten comments
* Documentation update (Codacy)
* Typos.
* Fix issues with online documentation
  Travis: switch from trusty to xenial to get more recent doxygen and get better markdown support
  CMake: remove commands to copy missing folder Docs/images
  Add badges

Code cleaning:
* Reformatting to match clang-tidy remarks
* Remove compilation warnings
* Remove unnecessary include.
* Use coherent argument name.
* Use bracket initializer instead of =.
* Uses initializer in constructors
* Missing lights in renderer : fixed
* attempt to clarify Camera and viewing parameters.
* Claryfing light and camera managers + small fixes
* Small rewriting in Shader building
* Remove pending texture when add the a tex for the same semantic.

+++ src/Engine/Renderer/RenderObject/RenderObjectManager.hpp
@@ -27,33 +27,55 @@ class RA_ENGINE_API RenderObjectManager final {
     Core::Index addRenderObject( RenderObject* renderObject );
     void removeRenderObject( const Core::Index& index );
 
-    uint getRenderObjectsCount();
+    size_t getRenderObjectsCount();
 
     /// Returns the render object corresponding to the given index. Will assert
     /// if the index does not match to an existing render object. See exists()
     std::shared_ptr<RenderObject> getRenderObject( const Core::Index& index );
 
     /**
-     * @brief Get all render objects, the vector is assumed to be empty
-     * @param Empty vector that will receive render objects
+     * @brief Get all render objects, the vector is assumed to be empty when called
+     * @param objectsOut vector that will receive render objects
      */
     void getRenderObjects( std::vector<std::shared_ptr<RenderObject>>& objectsOut ) const;
 
+    /**
+     * Get all render objects of the given type, the vector is assumed to be empty whan called
+     * @param objectsOut
+     * @param type
+     */
     void getRenderObjectsByType( std::vector<std::shared_ptr<RenderObject>>& objectsOut,
                                  const RenderObjectType& type ) const;
 
-    /// Returns true if the index points to a valid render object.
+    /** Returns true if the index points to a valid render object.
+     *
+     * @param index
+     * @return
+     */
     bool exists( const Core::Index& index ) const;
 
+    /**
+     * Removed the render object at the given index
+     * @param idx
+     */
     void renderObjectExpired( const Ra::Core::Index& idx );
 
-    /// Return the total number of faces drawn
-    uint getNumFaces() const;
+    /** Return the total number of faces drawn
+     *
+     * @return
+     */
+    size_t getNumFaces() const;
 
-    /// Return the total number of vertices drawn
-    uint getNumVertices() const;
+    /** Return the total number of vertices drawn
+     *
+     * @return
+     */
+    size_t getNumVertices() const;
 
-    /// Return the AABB of all visible render objects
+    /** Return the AABB of all visible render objects
+     *
+     * @return
+     */
     Core::Aabb getSceneAabb() const;
 
   private:
